requirements.txt
beautifulsoup4==4.12.2
Django==4.2.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pandas==2.1.1
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
rest-framework-simplejwt==0.0.2
